more careful handling of node indexes in highlighting code
The previous version worked fine in modern browsers but led to infinite recursion and an out of memory error in IE8.

The reason seems to be that inserting additional nodes changes the set for-in loops over. Fix that by using a »proper« for loop with an index and accounting for the additional nodes we insert.