Offer a remove Facets link when a facet is selected but does not appear
When metadata normalisation is not perfectly aligned with the
indexes it can happen that a facet is displayed but will not
yield any results (e.g. when metadata normalisation extracts all
language codes for an item but only the first one is indexed).

Instead of displaying an empty facet, now display a link to undo
the facet selection to give users a way out of this situation
without a complete reload.